Woodblock translation in 3d
###Hello magical people,
I am exploring this side of technology trying to help my office have a little fun in this new world.

We work on physical wood blocks (usually colored to represent the use of the room) at the beginning of the architecture design. After the initial design, we have to translate it into 3d software (Rhinoceros). An example image is in the reference image.
The original process of digitizing is pretty slow. I am imagining I can set up a camera and the camera can look at the physical woodblock then live-generate a 3d model in the digital environment. That will be awesome.

I am fairly unfamiliar with the software developer side; I appreciate it if anyone would like to share your input!###